Clase 71 - MAPEANDO ENTIDADES PARTNERS ADEMAS AGREGANDO DUDAS SOBRE EL METODO findByClientId 👤👤🕵

+++ src/main/java/com/george/springsecurity/repositories/PartnerRepository.java
@@ -3,10 +3,13 @@ package com.george.springsecurity.repositories;
 import com.george.springsecurity.entities.PartnerEntity;
 import org.springframework.data.repository.CrudRepository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 
-public interface PartnerRepository extends CrudRepository<PartnerEntity, Long> {
+public interface PartnerRepository extends CrudRepository<PartnerEntity, BigInteger> {
 
     Optional<PartnerEntity> findByClientId(String clientId);
+
+
 }
